Remove 'parsed' from tools names + add reddit tool
