+++ README.md
@@ -104,8 +104,8 @@ When no toolsets are specified, all tools are registered.
 | Toolset        | Tools                               |
 | -------------- | ----------------------------------- |
 | `web`          | `scrape_as_markdown`, `screenshot`, |
-| `search`       | `google_search_parsed`              |
-| `ecommerce`    | `amazon_search_parsed`              |
+| `search`       | `google_search`                     |
+| `ecommerce`    | `amazon_search`                     |
 | `social_media` | `reddit_post`, `reddit_subreddit`   |
 | `ai`           | `chatgpt`, `perplexity`             |
 
@@ -113,16 +113,16 @@ When no toolsets are specified, all tools are registered.
 
 The server exposes the following tools:
 
-| Tool                   | Description                                                                                | Example prompt                                                                          |
-| ---------------------- | ------------------------------------------------------------------------------------------ | --------------------------------------------------------------------------------------- |
-| `scrape_as_markdown`   | Scrapes any target URL, expects a URL to be given via prompt. Returns results in Markdown. | Scrape peacock.com from a US IP address and tell me the pricing.                        |
-| `screenshot`           | Captures a screenshot of any webpage and returns it as a PNG image.                        | Take a screenshot of github.com from a US IP address.                                   |
-| `google_search_parsed` | Scrapes Google Search for a given query, and returns parsed results.                       | Scrape Google Search for shoes and tell me the top position.                            |
-| `amazon_search_parsed` | Scrapes Amazon Search for a given query, and returns parsed results.                       | Scrape Amazon Search for toothbrushes.                                                  |
-| `reddit_post`          | Scrapes a specific Reddit post for a given query, and returns parsed results.              | Scrape the following Reddit post: https://www.reddit.com/r/horseracing/comments/1nsrn3/ |
-| `reddit_subreddit`     | Scrapes a specific Reddit subreddit for a given query, and returns parsed results.         | Scrape the top 5 posts on r/Python this week.                                           |
-| `chatgpt`              | Search and interact with ChatGPT for AI-powered responses and conversations.               | Ask ChatGPT to explain quantum computing in simple terms.                               |
-| `perplexity`           | Search and interact with Perplexity for AI-powered responses and conversations.            | Ask Perplexity what the latest trends in web development are.                           |
+| Tool                 | Description                                                                                | Example prompt                                                                          |
+| -------------------- | ------------------------------------------------------------------------------------------ | --------------------------------------------------------------------------------------- |
+| `scrape_as_markdown` | Scrapes any target URL, expects a URL to be given via prompt. Returns results in Markdown. | Scrape peacock.com from a US IP address and tell me the pricing.                        |
+| `screenshot`         | Captures a screenshot of any webpage and returns it as a PNG image.                        | Take a screenshot of github.com from a US IP address.                                   |
+| `google_search`      | Scrapes Google Search for a given query, and returns parsed results.                       | Scrape Google Search for shoes and tell me the top position.                            |
+| `amazon_search`      | Scrapes Amazon Search for a given query, and returns parsed results.                       | Scrape Amazon Search for toothbrushes.                                                  |
+| `reddit_post`        | Scrapes a specific Reddit post for a given query, and returns parsed results.              | Scrape the following Reddit post: https://www.reddit.com/r/horseracing/comments/1nsrn3/ |
+| `reddit_subreddit`   | Scrapes a specific Reddit subreddit for a given query, and returns parsed results.         | Scrape the top 5 posts on r/Python this week.                                           |
+| `chatgpt`            | Search and interact with ChatGPT for AI-powered responses and conversations.               | Ask ChatGPT to explain quantum computing in simple terms.                               |
+| `perplexity`         | Search and interact with Perplexity for AI-powered responses and conversations.            | Ask Perplexity what the latest trends in web development are.                           |
 
 ## Parameters
