--- src/tools/amazon-search-parsed/index.ts
@@ -1,1 +1,0 @@
-export * from './amazon-search-parsed-tool';
