+++ src/tools/amazon-search/index.ts
@@ -1,0 +1,1 @@
+export * from './amazon-search-tool';
