--- src/tools/google-search-parsed/index.ts
@@ -1,1 +1,0 @@
-export * from './google-search-parsed-tool';
