+++ src/tools/google-search/index.ts
@@ -1,0 +1,1 @@
+export * from './google-search-tool';
