+++ src/tools/index.ts
@@ -1,8 +1,9 @@
-export * from './amazon-search-parsed';
+export * from './amazon-search';
 export * from './chatgpt';
-export * from './google-search-parsed';
+export * from './google-search';
 export * from './perplexity';
 export * from './reddit-post';
 export * from './reddit-subreddit';
+export * from './reddit-user';
 export * from './scrape-as-markdown';
 export * from './screenshot';
